requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
pytz
django-filter>=23.3
pytest
pytest-cov
pytest-django